requirements.txt
fastapi==0.104.1
strawberry-graphql[fastapi]==0.212.0
pandas==2.1.3
uvicorn==0.24.0
python-multipart==0.0.6 